libs/tempfile: Move into internal
